// src/oracle/xrpl/models/methods/subscribe.ts
interface BaseStream {
  type: string
}
// ...
export interface TransactionStream extends BaseStream {
  type : "transaction"
  status: string
  /** String Transaction result code. */
  engine_result: string
  /** Numeric transaction response code, if applicable. */
  engine_result_code: number
  /** Human-readable explanation for the transaction response. */
  engine_result_message: string
  /**
   * The ledger index of the current in-progress ledger version for which this
   * transaction is currently proposed.
   */
  ledger_current_index?: number
  /** The identifying hash of the ledger version that includes this transaction. */
  ledger_hash?: string
  /** The ledger index of the ledger version that includes this transaction. */
  ledger_index?: number
  /**
   * The transaction metadata, which shows the exact outcome of the transaction
   * in detail.
   */
  meta?: TransactionMetadata
  /** The definition of the transaction in JSON format. */
  transaction: Transaction
  /**
   * If true, this transaction is included in a validated ledger and its
   * outcome is final. Responses from the transaction stream should always be
   * validated.
   */
  validated?: boolean
  warnings?: Array<{ id: number; message: string }>
}